+++ inc/ti/fn/fnnewbackup.h
@@ -9,13 +9,15 @@ static int do__f_new_backup(ti_query_t * query, cleri_node_t * nd, ex_t * e)
     uint64_t backup_id;
     uint64_t timestamp = 0;
     uint64_t repeat = 0;
+    uint64_t max_files = 1;
     ti_backup_t * backup;
     ti_raw_t * tar_gz_str = (ti_raw_t *) ti_val_borrow_tar_gz_str();
+    queue_t * files_queue;
 
     if (fn_not_node_scope("new_backup", query, e) ||
         ti_access_check_err(ti.access_node,
             query->user, TI_AUTH_MODIFY, e) ||
-        fn_nargs_range("new_backup", DOC_NEW_BACKUP, 1, 3, nargs, e) ||
+        fn_nargs_range("new_backup", DOC_NEW_BACKUP, 1, 4, nargs, e) ||
         ti_do_statement(query, child->node, e) ||
         fn_arg_str("new_backup", DOC_NEW_BACKUP, 1, query->rval, e))
         return e->nr;
@@ -33,6 +35,8 @@ static int do__f_new_backup(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     if (nargs >= 2)
     {
+        max_files = TI_BACKUP_DEFAULT_MAX_FILES;
+
         if (ti_do_statement(query, (child = child->next->next)->node, e))
             goto fail0;
 
@@ -77,7 +81,7 @@ static int do__f_new_backup(ti_query_t * query, cleri_node_t * nd, ex_t * e)
         query->rval = NULL;
     }
 
-    if (nargs == 3)
+    if (nargs >= 3)
     {
         if (ti_do_statement(query, (child = child->next->next)->node, e) ||
             fn_arg_int("new_backup", DOC_NEW_BACKUP, 3, query->rval, e))
@@ -89,7 +93,26 @@ static int do__f_new_backup(ti_query_t * query, cleri_node_t * nd, ex_t * e)
         query->rval = NULL;
     }
 
+    if (nargs == 4)
+    {
+        if (ti_do_statement(query, (child = child->next->next)->node, e) ||
+            fn_arg_int("new_backup", DOC_NEW_BACKUP, 4, query->rval, e))
+            goto fail0;
+
+        max_files = VINT(query->rval) < 1 ? 1 : (uint64_t) VINT(query->rval);
+
+        ti_val_unsafe_drop(query->rval);
+        query->rval = NULL;
+    }
+
     backup_id = ti.backups->next_id;
+
+    files_queue = queue_new(max_files);
+    if (!files_queue)
+    {
+        ex_set_mem(e);
+        goto fail0;
+    }
 
     backup = ti_backups_new_backup(
             backup_id,
@@ -97,10 +120,13 @@ static int do__f_new_backup(ti_query_t * query, cleri_node_t * nd, ex_t * e)
             rname->n,
             timestamp,
             repeat,
-            util_now_tsec());
+            max_files,
+            util_now_tsec(),
+            files_queue);
 
     if (!backup)
     {
+        free(files_queue);  /* still empty so we can simply use free */
         ex_set_mem(e);
         goto fail0;
     }