+++ inc/util/fx.h
@@ -23,6 +23,7 @@ unsigned char * fx_read(const char * fn, ssize_t * size);
 _Bool fx_file_exist(const char * fn);
 _Bool fx_is_dir(const char * path);
 int fx_mkdir_n(const char * path, size_t n);
+int fx_unlink_n(const char * str, size_t n);
 int fx_rmdir(const char * path);
 char * fx_path_join(const char * s1, const char * s2);
 char * fx_get_exec_path(void);